Added side running, made diamonds generate
Working on generation of rare world features

+++ Assets/Scripts/Objects/Biome.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public class Biome
 {
     public string Name;
-    public List<(string, int)> Features;
+    public List<(string, float)> Features;
 }
